Fix issue with item selection
e.target wasn’t returning the list element when clicking on the image
or text, so made it traverse up the tree until it finds the list
element. Possibly will cause an infinite while if it can’t find the
list but can’t think of a situation where this would occur.