Loading indicator, bugfix
Added loading indicator upon opening wads.
Minor cleanups
Fixed an issue upon loading very large wads (or wads with very large
lumps).

The engine is very conservative when loading wads to avoid unnecessary
errors like the one that arose. The problem was caused in the
interface, where it would check the type of every lump to provide the
correct icon. Unfortunately, one of the steps in lump type detection
was to load the entire lump and read just a few bytes at the start.
When attempting to do this on many large lumps at once, it would cause
the browser to lock. This was fixed by changing the way lumps are read
in detection by reading just the necessary bytes instead of getting the
entire lump.