Music detection for mapdata.js
Returns (currently a guess at) the music lump name for the map. This is
based on the hardcoded values for Doom, Doom 2 and Heretic. Most
vanilla PWADs, and a lot of wads without DEHACKED or MAPINFO lumps will
use the vanilla names for music files, so this will still be accurate.
It only populates a value if the lump exists in the wad.