Fix for ZDoom in Hexen formats
So ZDoom has many optional lumps when reading both Doom and Hexen
format maps. All of the nodebuilder-related lumps are optional, and
there is also an optional ZNODES lump for GLNodes. This means that it
is impossible to rely on the lump order of the map data, so I removed
that reliance.