mus2midi
this was a pain to do but i did it.

i rewrote the mus2midi.cpp found in slade into javascript. conversion
takes around 3/4s of a second on my machine/browser, not sure of how to
optimise it. Having some way of making it asynchronous would be nice i
guess